chore: keep external chain separated from eth utils
