+++ packages/nextjs/app/blockexplorer/page.tsx
@@ -5,8 +5,8 @@ import { PaginationButton, TransactionsTable } from "./_components";
 import type { NextPage } from "next";
 import { useFetchBlocks } from "~~/hooks/scaffold-eth";
 import { useTargetNetwork } from "~~/hooks/scaffold-eth/useTargetNetwork";
+import { arbitrumNitro } from "~~/utils/chain";
 import { notification } from "~~/utils/scaffold-eth";
-import { arbitrumNitro } from "~~/utils/scaffold-eth/chain";
 
 const BlockExplorer: NextPage = () => {
   const { blocks, transactionReceipts, currentPage, totalBlocks, setCurrentPage, error } = useFetchBlocks();
