+++ packages/nextjs/components/Footer.tsx
@@ -7,7 +7,7 @@ import { SwitchTheme } from "~~/components/SwitchTheme";
 import { Faucet } from "~~/components/scaffold-eth";
 import { useTargetNetwork } from "~~/hooks/scaffold-eth/useTargetNetwork";
 import { useGlobalState } from "~~/services/store/store";
-import { arbitrumNitro } from "~~/utils/scaffold-eth/chain";
+import { arbitrumNitro } from "~~/utils/chain";
 
 /**
  * Site footer
