+++ packages/nextjs/components/scaffold-eth/Address.tsx
@@ -8,8 +8,8 @@ import { useEnsAvatar, useEnsName } from "wagmi";
 import { CheckCircleIcon, DocumentDuplicateIcon } from "@heroicons/react/24/outline";
 import { BlockieAvatar } from "~~/components/scaffold-eth";
 import { useTargetNetwork } from "~~/hooks/scaffold-eth/useTargetNetwork";
+import { arbitrumNitro } from "~~/utils/chain";
 import { getBlockExplorerAddressLink } from "~~/utils/scaffold-eth";
-import { arbitrumNitro } from "~~/utils/scaffold-eth/chain";
 
 type AddressProps = {
   address?: AddressType;
