+++ packages/nextjs/components/scaffold-eth/Faucet.tsx
@@ -7,7 +7,7 @@ import { useNetwork } from "wagmi";
 import { BanknotesIcon } from "@heroicons/react/24/outline";
 import { Address, AddressInput, Balance, EtherInput } from "~~/components/scaffold-eth";
 import { useTransactor } from "~~/hooks/scaffold-eth";
-import { arbitrumNitro } from "~~/utils/scaffold-eth/chain";
+import { arbitrumNitro } from "~~/utils/chain";
 
 const localWalletClient = createWalletClient({
   account: privateKeyToAccount(arbitrumNitro.accounts[0].privateKey),
