+++ packages/nextjs/components/scaffold-eth/RainbowKitCustomConnectButton/BurnerWalletModal.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { arbitrumNitro } from "~~/utils/scaffold-eth/chain";
+import { arbitrumNitro } from "~~/utils/chain";
 
 interface BurnerWalletModalProps {
   isOpen: boolean;
